Project/requirements.txt
Django==3.2.8
psycopg2-binary==2.8.6
djangorestframework==3.12.4
django-cors-headers==3.10.0
drf-yasg==1.20.0
djangorestframework-simplejwt==5.0.0
django-rest-auth==0.9.5